% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvNet_Import_general.R
\name{add_values}
\alias{add_values}
\title{Add values}
\usage{
add_values(df)
}
\value{
A data frame with columns, where value labels have been added for specific variables
}
\description{
Add values
}
\examples{
add_values(import_SurvNet_general(build_query()))


}
