% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvNet_Import_general.R
\name{build_query}
\alias{build_query}
\title{Build query}
\usage{
build_query(TagKon = FALSE)
}
\arguments{
\item{TagKon}{}
}
\value{
a query for survNet
}
\description{
Build query
}
\examples{
query <- build_query(TagKon=TRUE)

}
