% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupIT.R
\name{check_r_storage_paths}
\alias{check_r_storage_paths}
\title{check_r_storage_paths}
\usage{
check_r_storage_paths()
}
\value{
writes R path configuration to console
}
\description{
check_r_storage_paths
}
\examples{
check_r_storage_paths()
}
