% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_pipeline_evaluation.R
\name{combine_assembly_stats}
\alias{combine_assembly_stats}
\title{combine_assembly_stats}
\usage{
combine_assembly_stats(old, new)
}
\arguments{
\item{old}{The dataframes derived from old RIDOM comparison tables}

\item{new}{The dataframes derived from new RIDOM comparison tables}
}
\value{
A dataframe showing assembly comparisons. The difference of the N50
values of old vs. new assemblies are given percent.
}
\description{
combine_assembly_stats
}
