% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_pipeline_evaluation.R
\name{combine_ehec_stats}
\alias{combine_ehec_stats}
\title{combine_ehec_stats}
\usage{
combine_ehec_stats(old, new)
}
\arguments{
\item{old}{The dataframes derived from old RIDOM comparison tables}

\item{new}{The dataframes derived from new RIDOM comparison tables}
}
\value{
a dataframe with comparisons for EHEC
}
\description{
combine_ehec_stats
}
