% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_pipeline_evaluation.R
\name{combine_general_stats}
\alias{combine_general_stats}
\title{combine_general_stats}
\usage{
combine_general_stats(old, new, what)
}
\arguments{
\item{old}{The dataframes derived from old RIDOM comparison tables}

\item{new}{The dataframes derived from new RIDOM comparison tables}

\item{what}{either Listeria, MRSA or Legionella}
}
\value{
a data frame containg comparison statistics for what
}
\description{
combine_general_stats
}
