% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_pipeline_evaluation.R
\name{combine_salmonella_stats}
\alias{combine_salmonella_stats}
\title{combine_salmonella_stats}
\usage{
combine_salmonella_stats(old, new)
}
\arguments{
\item{old}{The dataframes derived from old RIDOM comparison tables}

\item{new}{The dataframes derived from new RIDOM comparison tables}
}
\value{
a dataframe with comparisons for Salmonella
}
\description{
combine_salmonella_stats
}
