% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_pipeline_evaluation.R
\name{create_QMpdf_for_signatures}
\alias{create_QMpdf_for_signatures}
\title{create_QMpdf_for_signatures}
\usage{
create_QMpdf_for_signatures(
  qm_list,
  output_pdf = paste0("QM_pipeline_", format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M-\%S"),
    ".pdf"),
  temp_rmd = tempfile(fileext = ".Rmd")
)
}
\arguments{
\item{qm_list}{a list of tibbles produced by tabulate_pipeline_QM}

\item{output_pdf}{the output pdf (path) produced by the function (defaults to
"QM_pipeline_", a timestamp, and the ending ".pdf".}

\item{temp_rmd}{a temporary file to store the .Rmd (code) file}
}
\description{
Creates a pdf summarizing the comparison for the QM comparison and evaluation
of bioinformatic pipelines
}
