% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvNet_Import.R
\name{.append_GA_Kuerzel}
\alias{.append_GA_Kuerzel}
\title{.append_GA_Kuerzel}
\usage{
.append_GA_Kuerzel(data)
}
\arguments{
\item{data}{Usually a tibble generated by \code{\link{import_SurvNet()}}}
}
\value{
a tibble as generated by \code{\link{import_SurvNet()}}, but with
Meldelandkreis and their IDs included
}
\description{
Internal function (not exported) to add the ID (Kuerzel) of the
Gesundheitsamt of a particular Landkries (LK)
}
