% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_SurvNet_Import.R
\name{.append_LK_LatLong}
\alias{.append_LK_LatLong}
\title{.append_LK_LatLong}
\usage{
.append_LK_LatLong(data)
}
\arguments{
\item{data}{Usually a tibble generated by \code{\link{import_SurvNet()}}}
}
\value{
a tibble as generated by \code{\link{import_SurvNet()}}, but with
centroids (latitude and longitude) of the Landkreis included
}
\description{
Internal function (not exported), adds the latitude and longitude of the
centroid of each Landkreis, see \code{\link{LKLatLong}}
}
