% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_SurvNet_Import.R
\name{.cleanup_RIDOM}
\alias{.cleanup_RIDOM}
\title{.cleanup_RIDOM}
\usage{
.cleanup_RIDOM(data)
}
\arguments{
\item{data}{Usually a tibble generated by \code{\link{import_SurvNet()}} and
data for Gesundheitsaemter and Landkreise appended}
}
\value{
a tibble
}
\description{
Internal function (not exported), which cleans and prepares SurvNet data
for import into RIDOM according to the database scheme defined there.
}
