% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvNet_Import.R
\name{.combine_exposure}
\alias{.combine_exposure}
\title{.combine_exposure}
\usage{
.combine_exposure(data)
}
\arguments{
\item{data}{Usually a tibble generated by \code{\link{import_SurvNet()}}}
}
\value{
A tibble derived the
tibble as generated by \code{\link{import_SurvNet()}}, but with
Expositionsort standardized
}
\description{
.combine_exposure
}
