% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Eigentuemer}
\alias{Eigentuemer}
\title{Gesundheitsaemter and their IDs: Eigentuemer}
\format{
A tibble with 24 rows and 3 variables:
\describe{
  \item{Eigentuemer}{ID for the Gesunheitsamt}
  \item{Gesundheitsam}{Name (Ortsname) of the Gesundheitsamt}
  \item{Kuerzel}{Short abbreviation for the Gesundheitsamt derived from
  location identifier used on car license plates}
}
}
\source{
Compiled by Anja Schoeps
}
\usage{
data(Eigentuemer)
}
\description{
Gesundheitsaemter "own" (German: "sind die Eigetuemer") of the data in
SurvNet, this dataset contains information on how we use them assigning
IDs to and from them.
}
\examples{
data(Eigentuemer)
head(Eigentuemer)
}
\keyword{datasets}
