% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Einlesen_und_Aufbereiten_SURE.R
\name{einlesen_sure}
\alias{einlesen_sure}
\title{einlesen_sure}
\usage{
einlesen_sure(
  path = "Z:/DFS-LUA-LD-Zusammenarbeit/LD-AB32.5_IfSG_Meldewesen/SURE/Daten"
)
}
\arguments{
\item{path}{Pfad auf dem der LIMS-Export liegt}
}
\value{
Ein datenframe zur Weiterverwendung
}
\description{
einlesen_sure
}
\examples{
einlesen_sure()
}
