% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_SurvNet_Import.R
\name{geo2ridom}
\alias{geo2ridom}
\title{geo2ridom}
\usage{
geo2ridom(x)
}
\arguments{
\item{x}{a POINT geometry from sf}
}
\value{
a RIDOM compatible geo-coordinate point
}
\description{
geo2ridom
}
