% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{IdType_Datensatzkategorie}
\alias{IdType_Datensatzkategorie}
\title{IdType_Datensatzkategorie}
\format{
Eine Tabelle mit 2 Spalten:
\describe{
  \item{IdType}{Vom RKI benutzte Kodierung (3 Ziffern)}
  \item{Datensatzkategorie}{ausformulierte Meldekategorie}
  }
}
\source{
Compiled by Anja Schoeps (from SurvNet)
}
\usage{
data(IdType_Datensatzkategorie)
}
\description{
Eine Tabelle mit 2 Spalten, die dem Code für IdType in SurvNet die
korrekte Bezeichnung der Meldekategorie zuordnet
}
\examples{
data(IdType_Datensatzkategorie)
}
\keyword{datasets}
