% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvNet_Import_general.R
\name{import_SurvNet_general}
\alias{import_SurvNet_general}
\title{import_SurvNet}
\usage{
import_SurvNet_general(query = build_query())
}
\value{
A data frame specified columns (hardcoded for now)
}
\description{
import_SurvNet
}
\examples{
import_SurvNet_general(build_query())


}
