% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_count_lines.R
\name{launch_count_lines}
\alias{launch_count_lines}
\title{Launch the Line Counter Shiny App}
\usage{
launch_count_lines()
}
\description{
This function starts the Shiny app for counting lines in a file.
}
