% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_qPCR_evaluator.R
\name{launch_qPCR_evaluator}
\alias{launch_qPCR_evaluator}
\title{Launch the qPCR Evaluator Shiny App}
\usage{
launch_qPCR_evaluator()
}
\description{
This function starts the Shiny app for Evaluating qPCR results.
}
