% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_ridom_2_lims.R
\name{launch_ridom_2_lims}
\alias{launch_ridom_2_lims}
\title{Launch the Ridom 2 LIMS Shiny App}
\usage{
launch_ridom_2_lims()
}
\description{
This function starts the Shiny app for preparing RIDOM outputs for LIMS
import.
}
