% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupIT.R
\name{lua_is_online}
\alias{lua_is_online}
\title{lua_is_online Überprüfe ob R online-Zugang hat}
\usage{
lua_is_online(timeout = 15, url = "http://www.google.com")
}
\arguments{
\item{timeout}{}
}
\value{
logical indicating whether you're online
}
\description{
Die Funktion checkt ob innerhalb einer gewissen Latenzzeit (timeout) eine
Website erreicht werden kann. Dies ist besonders nützlich um bei
Fehlermeldugnen während der Paketinstallation zu prüfen, ob die Proxy-
Einstellungen in R korrekt sind und eine Netzwerkverbindung zulassen
}
\examples{
lua_is_online()

}
