% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupIT.R
\name{lua_network_speedtest}
\alias{lua_network_speedtest}
\title{lua_network_speedtest()}
\usage{
lua_network_speedtest(
  net_path = "O:/Abteilung Humanmedizin (AHM)/Referat 32/R_scratch/R_IO_test_network.txt",
  loc_path = file.path(Sys.getenv("USERPROFILE"), "Desktop", "R_IO_test_local.txt")
)
}
\arguments{
\item{net_path}{Path on the Network you want to test. Defaults to a path on O:}

\item{loc_path}{Local path. Defaults to a path on your user's desktop}
}
\value{
writes speed estimates to console
}
\description{
lua_network_speedtest()
}
\examples{
lua_network_speedtest()
}
