% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{pathogen_dictionary}
\alias{pathogen_dictionary}
\title{pathogen_dictionary}
\format{
Eine Tabelle mit 2 Spalten:
\describe{
  \item{Pathogen_ID}{Vom RKI benutzte Kodierung (gewöhnlich 7 Ziffern)}
  \item{Erreger}{ausformulierter Erreger/Serovar}
  }
}
\source{
Compiled by Anja Schoeps
}
\usage{
data(pathogen_dictionary)
}
\description{
Eine Tabelle mit 2 Spalten, die dem Code für die Pathogene in SurvNet die
korrekte Bezeichnung des Erregers/Serovars zuordnet.
Diese Datei wird von der Funktion add_values() benutzt.
}
\examples{
data(pathogen_dictionary)
}
\keyword{datasets}
