% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Praxisliste}
\alias{Praxisliste}
\title{Praxisliste}
\format{
A list with two geo-objects (Sf - Simple feature collections):
\describe{
  \item{einsender}{foo}
  \item{praxisname}{foo}
  }
}
\source{
Compiled by Anja Schoeps
}
\usage{
data(Praxislist)
}
\description{
Foo
}
\examples{
data(Praxisliste)
}
\keyword{datasets}
