% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Regionen}
\alias{Regionen}
\title{Regionen}
\format{
Eine Tabelle mit 2 Spaltn:
\describe{
  \item{ID_Region}{Vom RKI benutzte Kodierung (gewöhnlich 8 Ziffern)}
  \item{Angabe_zum_Expositionsort}{ausformulierter Expositionsort}
  }
}
\source{
Compiled by Anja Schoeps
}
\usage{
data(Regionen)
}
\description{
Eine Tabelle mit 2 Spalten, die dem Code für Region_ID in SurvNet die
korrekte Bezeichnung der Region zuordnet. Diese Variable wird für die
Zuordnung der Expositionsorte benutzt
}
\examples{
data(Regionen)
}
\keyword{datasets}
