% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_pipeline_evaluation.R
\name{tabulate_pipeline_QM}
\alias{tabulate_pipeline_QM}
\title{tabulate_pipeline_QM}
\usage{
tabulate_pipeline_QM(x)
}
\arguments{
\item{a}{list of tibbles from RIDOM_comptable_re_evaluate}
}
\value{
a list of summarized tibbles comparing the new stats vs. the old
}
\description{
tabulate_pipeline_QM
}
