% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{themes_list}
\alias{themes_list}
\title{themes_list}
\format{
Eine Tabelle mit 4 Spalten, mit folgenden SQL-Blöcken :
\describe{
  \item{Thema}{Name des Themas}
  \item{Select}{Passender Select-Block für das jeweilige Thema}
  \item{From}{Passender From-Block für das jeweilige Thema}
  \item{Where}{Passender Where-Block für das jeweilige Thema}
  }
}
\source{
Compiled by Anja Schoeps
}
\usage{
data(themes_list)
}
\description{
Liste der Themen, die standardmäßig aus SurvNet abgefragt werden können
}
\details{
Jedem Themenbereich, der mittels \code{\link{import_SurvNet}} standardmäßig
abgefragt werden kann, sind die passenden SQL-Blöcke zugeordnet. Diese SQL-Blöcke
werden in \code{\link{build_query}} verknüpft, um die finale Abfrage zu generieren.

Dieser Datensatz enthält alle Themen, die in \code{build_query} angegeben werden können.
Jeder Eintrag besteht aus einem Themenbereich und den zugehörigen SQL-Blöcken,
die für die Abfrage verwendet werden.
}
\examples{
data(themes_list)
}
\keyword{datasets}
