% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Einlesen_und_Aufbereiten_SURE.R
\name{Zusammenfassung_Berichtswoche}
\alias{Zusammenfassung_Berichtswoche}
\title{Title}
\usage{
Zusammenfassung_Berichtswoche(sure_data)
}
\arguments{
\item{sure_data}{}
}
\value{
Einzeiliger Datenframe mit aktuellen Werten für die Berichte
}
\description{
Title
}
\examples{
Zusammenfassung_Berichtswoche()
}
