% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Einlesen_und_Aufbereiten_SURE.R
\name{aggregate_woche}
\alias{aggregate_woche}
\title{Title}
\usage{
aggregate_woche(sure_data, week, year)
}
\arguments{
\item{sure_data}{}
}
\value{
Aggregierter Datenframe mit ausgewählten Wochen
}
\description{
Title
}
\examples{
Zusammenfassung_Berichtswoche()

}
