% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvNet_Import_general.R
\name{build_query}
\alias{build_query}
\title{Funktion zum Aufbau eines Abfragetexts (SQL-Language) für die SurvNet-Abfrage}
\usage{
build_query(
  Thema = NULL,
  Periode = c("alle", "last2weeks", "thisyear", "last2years", "last5years",
    "last10years", "IMS_since_2023")
)
}
\arguments{
\item{Thema}{Auswahl zwischen alle, Tageskontrolle, VHF, WBK, Arboviren, ZoonotischeINV, SerovarSAL, Influenza, COVID19, RSV, IMS_RIDOM}

\item{Periode}{Auswahl zwischen alle, last2weeks, thisyear, last2years, last5years, last10years, IMS_since_2023}
}
\value{
Ausgabe wird in der Funktion import_SurvNet verwendet.
}
\description{
Es können spezielle vorgefertigte Abfragen erstellt werden.//
             Zusätzlich kann die Periode der Abfrage unabhängig davon festgelegt werden.
             Das Thema Tageskontrolle wird gewöhnlich mit der Periode last2weeks kombiniert
             Das Thema IMS_RIDOM wird gewöhnlich mit der Periode IMS_since_2023 kombiniert.
}
\details{
SurvNet_Abfrage
}
\examples{
query <- build_query(Thema="Tageskontrolle")

}
