% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvNet_Import.R
\name{create_Epidata}
\alias{create_Epidata}
\title{createEpidata}
\usage{
create_Epidata(
  LIMS_link_file,
  out = paste0("O:/Abteilung Humanmedizin (AHM)/Referat 32/32_6/14_EpiDaten/",
    "SurvNetExport4RIDOM", format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M-\%S"), ".csv")
)
}
\arguments{
\item{LIMS_link_file}{The file (exported from our LIMS) associating our
LUA Probennummern (3263-%) and SurvNet IDs (Aktenzeichen with licence plate
prefix)}

\item{out}{The output file in csv format (defaults to  SurvNetExport4RIDOM
plus the current timestamp wirtten into
"O:/Abteilung Humanmedizin (AHM)/Referat 32/32_6/14_EpiDaten/")}
}
\value{
Writes a file to out
}
\description{
createEpidata
}
