% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIDOM_SurvNet_Import.R, R/SurvNet_Import.R
\name{create_Epidata}
\alias{create_Epidata}
\title{createEpidata}
\format{
The xlsx file written to "out" has the following columns:
\describe{
  \item{Sample ID}{Character. As "Laborprobennummer" from LIMS}
  \item{Alias ID (s)}{Character. This is the \strong{merge key}! As
  "Survnet" from SurvNet and as "Aktenzeichen" from LIMS.}
  \item{Zip of Isolation}{Character. As Postleitzahl from LIMS}
  \item{IdRecord}{Character. Unchanged from SurvNet}
  \item{Collection Date}{Date. As Meldedatum from SurvNet}
  \item{Gesundheitsamt}{Character. Via merge based on
  Eigentuemer from SurvNet (merge key) obtaining the Gesundheitsamt
  variable from data("Eigentuemer")}
  \item{Meldelandkreis}{Character. Unchanged from SurvNet}
  \item{MunicipalityKey}{Character. Unchanged from SurvNet}
  \item{Host Age (years)}{Numeric. As AgeComputed from SurvNet}
  \item{Host Sex}{Character. Recoded from Geschlecht from SurvNet}
  \item{HospStatus}{Character. Recoded from HospitalisierungStatus from
  SurvNet}
  \item{VerstorbenStatus}{Character. Recoded from VerstorbenStatus from
  SurvNet}
  \item{Expositionsort}{Character. Derived from multiple exposure variables
  from SurvNet (see ?luaRlp:::.combine_exposure)}
  \item{Outbreak}{Logical. Recoded from SurvNet by determining
  whether AusbruchInfo_InterneRef is filled}
  \item{Country of Isolation}{Character. Hard-coded "Germany".}
  \item{Lat/Long of Isolation}{Character. The latitude and longitude of the
  centroid of the Landkreis derived from data("LKLatLong") merged via
  "Meldelandkreis" (merge key)}
  \item{Project}{Character. The RIDOM project name recoded from
  Meldekategorie (Erreger) from SurvNet}
  \item{Host}{Character. Hardcoded "homo sapiens" for all but Legionella,
  which needs to be entered manually (set to "" - empty)}
  \item{Source type}{Character. Hardcoded "clinical/host-associated" for all
  but Legionella, which needs to be entered manually (set to "" - empty)}
  \item{Source subtype}{Character. Hardcoded "human" for all but Legionella,
  which needs to be entered manually (set to "" - empty)}
  \item{Specimen type}{Character. Hardcoded "stool" for salmonella and EHEC,
  ""- empty for Legionella, MRSA and Listeria}
  }

The xlsx file written to "out" has the following columns:
\describe{
  \item{Sample ID}{Character. As "Laborprobennummer" from LIMS}
  \item{Alias ID (s)}{Character. This is the \strong{merge key}! As
  "Survnet" from SurvNet and as "Aktenzeichen" from LIMS.}
  \item{Zip of Isolation}{Character. As Postleitzahl from LIMS}
  \item{IdRecord}{Character. Unchanged from SurvNet}
  \item{Collection Date}{Date. As Meldedatum from SurvNet}
  \item{Gesundheitsamt}{Character. Via merge based on
  Eigentuemer from SurvNet (merge key) obtaining the Gesundheitsamt
  variable from data("Eigentuemer")}
  \item{Meldelandkreis}{Character. Unchanged from SurvNet}
  \item{MunicipalityKey}{Character. Unchanged from SurvNet}
  \item{Host Age (years)}{Numeric. As AgeComputed from SurvNet}
  \item{Host Sex}{Character. Recoded from Geschlecht from SurvNet}
  \item{HospStatus}{Character. Recoded from HospitalisierungStatus from
  SurvNet}
  \item{VerstorbenStatus}{Character. Recoded from VerstorbenStatus from
  SurvNet}
  \item{Expositionsort}{Character. Derived from multiple exposure variables
  from SurvNet (see ?luaRlp:::.combine_exposure)}
  \item{Outbreak}{Logical. Recoded from SurvNet by determining
  whether AusbruchInfo_InterneRef is filled}
  \item{Country of Isolation}{Character. Hard-coded "Germany".}
  \item{Lat/Long of Isolation}{Character. The latitude and longitude of the
  centroid of the Landkreis derived from data("LKLatLong") merged via
  "Meldelandkreis" (merge key)}
  \item{Project}{Character. The RIDOM project name recoded from
  Meldekategorie (Erreger) from SurvNet}
  \item{Host}{Character. Hardcoded "homo sapiens" for all but Legionella,
  which needs to be entered manually (set to "" - empty)}
  \item{Source type}{Character. Hardcoded "clinical/host-associated" for all
  but Legionella, which needs to be entered manually (set to "" - empty)}
  \item{Source subtype}{Character. Hardcoded "human" for all but Legionella,
  which needs to be entered manually (set to "" - empty)}
  \item{Specimen type}{Character. Hardcoded "stool" for salmonella and EHEC,
  ""- empty for Legionella, MRSA and Listeria}
  }
}
\usage{
create_Epidata(
  LIMS_link_file =
    "O:/Abteilung Humanmedizin (AHM)/Referat 32/32_6/14_EpiDaten/LIMS_Export_auto/LIMS_Export.csv",
  out =
    paste0("O:/Abteilung Humanmedizin (AHM)/Referat 32/32_6/14_EpiDaten/LIMS Import/",
    "SurvNetExport4RIDOM", format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M-\%S"), ".xlsx"),
  problems = NULL,
  type_of_problem = "non-matching-LIMS"
)

create_Epidata(
  LIMS_link_file =
    "O:/Abteilung Humanmedizin (AHM)/Referat 32/32_6/14_EpiDaten/LIMS_Export_auto/LIMS_Export.csv",
  out =
    paste0("O:/Abteilung Humanmedizin (AHM)/Referat 32/32_6/14_EpiDaten/LIMS Import/",
    "SurvNetExport4RIDOM", format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M-\%S"), ".xlsx"),
  problems = NULL,
  type_of_problem = "non-matching-LIMS"
)
}
\arguments{
\item{LIMS_link_file}{The file (exported from our LIMS) associating our
LUA Probennummern (3263-) and SurvNet IDs (Aktenzeichen with licence plate
prefix), defaults to a path on our System on which automated LIMS exports are
stored}

\item{out}{The output file in xlsx format (defaults to  SurvNetExport4RIDOM
plus the current timestamp wirtten into
"O:/Abteilung Humanmedizin (AHM)/Referat 32/32_6/14_EpiDaten/")}

\item{problems}{If problematic entries should be reported into a file
(a filename or path to a file) should be given. Which entries are regarded as
problematic are defined via type_of_problem}

\item{type_of_problem}{should be "non-matching-LIMS". The currently only
available option (also the current default) is to write non-matching LIMS
entries to a file (given in the argument "problems").}
}
\value{
Writes a xlsx file compatible with the RIDOM database scheme "LUA RLP
default bacteria to "\strong{out}" and a optionally a second csv file to
"\strong{problem}". See "Format".

Writes a xlsx file compatible with the RIDOM database scheme "LUA RLP
default bacteria to "\strong{out}" and a optionally a second csv file to
"\strong{problem}". See "Format".
}
\description{
createEpidata

createEpidata
}
