% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{geo_standards}
\alias{geo_standards}
\title{geo_standards}
\format{
Eine Tabelle mit 4 Spalten, in denen Informationen zu :
\describe{
  \item{Meldelandkreis}{LUA RLP interne Standard-Schreibweise der Landkreise}
  \item{GID_2}{Geo ID wie in der gadm Datenbank}
  \item{LD_ID}{Landkreis Identifyer, kompatibel mit SurvNet}
  \item{Gesundheitsamt}{Für den Landkreis zuständiges Gesundheitsamt}
  \item{Kürzel}{LUA RLP interes Standard-Kürzel für das Gesundheitsamt}
  \item{geom}{Polygone der Landkreis-Geometrie aus gadm}
  }
}
\source{
Compiled by Anja Schoeps with additions by Emanuel Heitlinger
}
\usage{
data(geo_standards)
}
\description{
Enthält eine Tabelle zur standardisierten Benennung und Verarbeitung von
Gebietsdaten mit Bezug zum öffentlichen Gesundheitsdienst in Rheinland-Pfalz
}
\examples{
data(geo_standards)
}
\keyword{datasets}
