% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Praxisliste}
\alias{Praxisliste}
\title{Praxisliste}
\format{
Eine Tabelle mit 5 Spalten, in denen zus채tzliche Informationen zu den Arztpraxen liegen:
\describe{
  \item{einsender}{LIMS-K체rzel f체r die Praxis}
  \item{praxisname}{Name der Arztpraxis}
  \item{Kreis}{Landkreis, in dem die Arztpraxis liegt}
  \item{Zuordnung}{LA oder KO: Labor, in dem Proben analysiert werden}
  \item{Arzttyp}{Kinder oder Allgemein}
  }
}
\source{
Compiled by Anja Schoeps
}
\usage{
data(Praxisliste)
}
\description{
Enh채lt Daten zu den teilnehmenden Arztpraxen in SURE
}
\examples{
data(Praxisliste)
}
\keyword{datasets}
